Move the firebase updates utils into a standalone module
